Adding twist to the interaction api
