+++ packages/viewer/src/interactions/mouseInteractions.ts
@@ -1,5 +1,5 @@
 import { InteractionApi } from './interactionApi';
-import { Point, Vector3, Angle } from '@vertexvis/geometry';
+import { Point } from '@vertexvis/geometry';
 
 export class MouseInteraction {
   public beginDrag(event: MouseEvent, api: InteractionApi): void {
@@ -153,8 +153,6 @@ export class PanInteraction extends MouseInteraction {
 export class TwistInteraction extends MouseInteraction {
   private currentPosition: Point.Point | undefined;
 
-  private lastAngle: number | undefined;
-
   public beginDrag(event: MouseEvent, api: InteractionApi): void {
     this.currentPosition = Point.create(event.screenX, event.screenY);
     api.beginInteraction();
@@ -162,19 +160,7 @@ export class TwistInteraction extends MouseInteraction {
 
   public drag(event: MouseEvent, api: InteractionApi): void {
     const position = Point.create(event.screenX, event.screenY);
-    api.transformCamera((camera, viewport) => {
-      const center = Point.create(viewport.width / 2, viewport.height / 2);
-      const currentAngle = Angle.fromPoints(center, position);
-      const angleDelta =
-        this.lastAngle != null ? currentAngle - this.lastAngle : 0;
-
-      this.lastAngle = currentAngle;
-      const axis = Vector3.normalize(
-        Vector3.subtract(camera.lookAt, camera.position)
-      );
-      const angleInRadians = Angle.toRadians(-angleDelta);
-      return camera.rotateAroundAxis(angleInRadians, axis);
-    });
+    api.twistCamera(position);
   }
 
   public endDrag(event: MouseEvent, api: InteractionApi): void {
